% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderShapeEM.R
\name{OrderShapeEM.control}
\alias{OrderShapeEM.control}
\title{Set the control parameters for OrderShapeEM}
\usage{
OrderShapeEM.control(
  maxIter = 250,
  tol = 0.001,
  trace = FALSE,
  pi0.init = 0.95,
  k.init = 0.75,
  pvals.cutoff = 1e-15
)
}
\arguments{
\item{maxIter}{an integer value indicating the maximum number of iterations. Default is 250.}

\item{tol}{a numeric value giving the tolerance in the relative change in the log likelihood below which the algorithm is considered to be converged.
Default is 1e-3.}

\item{trace}{a logical value indicating whether to print out the EM process. Default is FALSE.}

\item{pi0.init}{a scalar giving the initial estimates of pi0. Default is 0.95.}

\item{k.init}{a scalar giving the initial estimates of f1. f1 is approximated by a beta distribution with the shape parameter k. 
Default is 0.75.}

\item{pvals.cutoff}{a numeric value to replace p-values below that value, which is used to increase the stability of the algorithm. Default is 1e-15.}
}
\value{
A list with the set control parameters.
}
\description{
The function provides an interface to set the parameters for OrderShapeEM.
}
